airwriting-advanced/requirements.txt
# requirements.txt - AirWriting Advanced 기본 의존성
# Ubuntu 20.04+ 환경에서 테스트됨

# 수치 계산
numpy>=1.23.0
scipy>=1.10.0

# 설정 파일 관리
PyYAML>=6.0

# I2C 통신
smbus2>=0.4.3

# 시리얼 통신
pyserial>=3.5

# 수치 안정성
scikit-learn>=1.3.0

# 시각화 (기본)
matplotlib>=3.7.0
